+++ test/static_test.cpp
@@ -1,4 +1,4 @@
-#include "../cpp_dump.hpp"
+#include "../cpp-dump.hpp"
 #include "./odr_test.hpp"
 
 template <typename... Args>
